+++ packages/inqvine_core_simulate/lib/types/simulatable_go_route.dart
@@ -5,9 +5,11 @@ class SimulatableGoRoute {
     required this.name,
     required this.route,
     required this.fakeModel,
+    this.group = 'Ungroupped',
   });
 
   final String name;
+  final String group;
   final GoRoute route;
   final SimulationActionable fakeModel;
 }